Update README.md
This mini-project is designed to detect obstacles using ultrasonic sensors and automatically control motor operations to prevent collisions. It utilizes the LPC2148 microcontroller, which is based on the ARM7 TDMI architecture, making it suitable for embedded and real-time applications.

The primary objective is to simulate and implement an embedded crash prevention system that can be integrated into autonomous vehicles or robotic applications.